Add files via upload
Side Note: the attached 'Bosses document' is the old concepts sheet we did over a year ago 
                                                           THE REST OF THIS IS FOR ISRAEL
We are going to do what we were planning on doing originally, with the concepts but on a somewhat smaller scale. For now, I want to to start on level design so that you can get the tough stuff out of the way first. The level design should be based on Zelda 2/ plat-former design. 
https://www.google.ca/url?sa=i&rct=j&q=&esrc=s&source=images&cd=&cad=rja&uact=8&ved=0ahUKEwjW7uSy1ebMAhVK0oMKHThzAd4QjRwIBw&url=http%3A%2F%2Fwww.rpgfan.com%2Freviews%2Fzelda2%2FZelda_2-3.html&psig=AFQjCNEAJQwIYfwxJkyaYSh2RkyKvBPQRQ&ust=1463765452806011
https://www.google.ca/url?sa=i&rct=j&q=&esrc=s&source=images&cd=&cad=rja&uact=8&ved=&url=http%3A%2F%2Fwww.nesmaps.com%2Fmaps%2FZelda2%2FZelda2MapsTowns.html&psig=AFQjCNEAJQwIYfwxJkyaYSh2RkyKvBPQRQ&ust=1463765452806011
https://www.google.ca/url?sa=i&rct=j&q=&esrc=s&source=images&cd